+++ mage/src/main/java/mil/nga/giat/mage/event/EventActivity.kt
@@ -61,7 +61,7 @@ class EventActivity : DaggerAppCompatActivity() {
         }
 
         viewManager = LinearLayoutManager(this)
-        val forms = event?.forms?.map {
+        val forms = event?.forms?.mapNotNull {
             Form.fromJson(it?.asJsonObject)
         } ?: emptyList()